+++ samples/snippets/sendTemplateFreeformInvite.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const signnow = require('signnow')({
+const signnow = require('@signnow/api-client')({
   credentials: 'BASE64_ENCODED_CLIENT_CREDENTIALS',
   production: true, // if false then uses eval server
 });
@@ -8,7 +8,7 @@ const sendFreeformInvite = signnow.document.invite;
 
 const from = 'EMAIL_OF_SENDER_GOES_HERE';
 const to = 'EMAIL_OF_RECIPIENT_GOES_HERE';
-const id = 'DOCUMENT_ID_GOES_HERE';
+const id = 'TEMPLATE_ID_GOES_HERE';
 const token = 'YOUR_ACCESS_TOKEN';
 
 /**